chore(auth-server): Add fxaMailer mock setup
Because:
 - We have a lot of tests that will need a new mockFxaMailer
   implementation

This Commit:
 - Adds a shared mockFxaMailer setup function with ability to override
   properties for testing.

Closes: FXA-12933
